# [BE] Optimize last_successful_workflow ClickHouse query 