--- a/torchci/clickhouse_queries/last_successful_workflow/params.json
+++ b/torchci/clickhouse_queries/last_successful_workflow/params.json
@@ -2,5 +2,7 @@
   "params": {
     "workflowName": "String"
   },
-  "tests": []
+  "tests": [
+    "workflowName=docker-builds"
+  ]
 }