--- a/torchci/clickhouse_queries/last_successful_workflow/params.json
+++ b/torchci/clickhouse_queries/last_successful_workflow/params.json
@@ -3,6 +3,6 @@
     "workflowName": "String"
   },
   "tests": [
-    "workflowName=docker-builds"
+    {"workflowName": "docker-builds"}
   ]
 }