--- a/torchci/clickhouse_queries/last_successful_workflow/query.sql
+++ b/torchci/clickhouse_queries/last_successful_workflow/query.sql
@@ -1,19 +1,28 @@
 SELECT
-    DATE_DIFF(
-        'second',
-        workflow.created_at,
-        CURRENT_TIMESTAMP()
-    ) AS last_success_seconds_ago
+    DATE_DIFF('second', workflow.created_at, CURRENT_TIMESTAMP()) AS last_success_seconds_ago
 FROM
-    default.workflow_run workflow FINAL
-JOIN default.push FINAL ON workflow.head_commit.'id' = push.head_commit.'id'
-WHERE
-    push.ref IN ('refs/heads/master', 'refs/heads/main')
-    AND push.repository.'owner'.'name' = 'pytorch'
-    AND push.repository.'name' = 'pytorch'
-    AND workflow.conclusion = 'success'
-    AND workflow.name = {workflowName: String}
+(
+    SELECT
+      head_commit.id as head_commit_id,
+      created_at
+    FROM default.workflow_run
+    PREWHERE
+      conclusion = 'success'
+      AND created_at > CURRENT_TIMESTAMP() - INTERVAL 15 DAY
+      AND name = {workflowName: String}
+    ORDER BY created_at DESC
+) AS workflow
+JOIN
+(
+    SELECT
+      push.head_commit.id as head_commit_id
+    FROM default.push
+    PREWHERE
+      ref IN ('refs/heads/master', 'refs/heads/main')
+      AND repository.owner.name = 'pytorch'
+      AND repository.name = 'pytorch'
+) AS push
+ON workflow.head_commit_id = push.head_commit_id
 ORDER BY
     workflow.created_at DESC
-LIMIT
-    1
+LIMIT 1