@Camyll 2025-03-25
Approve to unblock, pls run sqlfluff linter